+++ sw/src/LayoutHost.cpp
@@ -5,7 +5,7 @@ void sw::LayoutHost::Associate(ILayout *obj)
     this->_associatedObj = obj;
 }
 
-uint32_t sw::LayoutHost::GetLayoutTag()
+uint64_t sw::LayoutHost::GetLayoutTag()
 {
     return this->_associatedObj->GetLayoutTag();
 }
